Refact app structure
