--- node/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "vtex.my-account",
-  "version": "0.14.0",
-  "main": "index.js"
-}
