--- node/service.json
@@ -1,10 +1,0 @@
-{
-  "stack": "nodejs",
-  "memory": 256,
-  "routes": {
-    "enabled": {
-      "path": "/_v/account/enabled",
-      "public": true
-    }
-  }
-}
