+++ react/components/loaders/AddressCreateLoading.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import BaseLoading from '../shared/BaseLoading'
-import AddressCreateHeader from './AddressCreateHeader'
-import SkeletonPiece from '../../components/shared/SkeletonPiece'
-import SkeletonBox from '../../components/shared/SkeletonBox'
+import AddressCreateHeader from '../headers/AddressCreateHeader'
+import SkeletonPiece from '../shared/SkeletonPiece'
+import SkeletonBox from '../shared/SkeletonBox'
 
 const AddressCreateLoading = ({ data }) => {
   return (
