+++ react/global.css
@@ -1,13 +1,3 @@
-/* Setting systemwide font to Helvetica */
-body,
-button,
-input,
-optgroup,
-select,
-textarea {
-  font-family: 'helvetica neue', helvetica, sans-serif;
-}
-
 /* Add tweakpoint inside m */
 @media screen and (min-width: 57em) {
   .dn-m-2 {
