+++ react/index.js
@@ -2,13 +2,15 @@ import React from 'react'
 import AppRouter from './components/AppRouter'
 import ClientSide from './components/ClientSide'
 import 'vtex.country-codes/locales'
-import './style.global.css'
+import './global.css'
 
 const bootstrap = () => {
   return (
-    <ClientSide>
-      <AppRouter />
-    </ClientSide>
+    <div className="vtex-account helvetica">
+      <ClientSide>
+        <AppRouter />
+      </ClientSide>
+    </div>
   )
 }
